+++ packages/charts/src/hooks/use-chart-axis/index.ts
@@ -1,0 +1,1 @@
+export { useChartAxis } from './use-chart-axis';