+++ postcss.config.cjs
@@ -1,0 +1,4 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+const { postcss } = require('@lualtek/config');
+
+module.exports = postcss.getConfig();