+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @lualtek/icons
 
+## 12.0.0-next.8
+
+## 12.0.0-next.7
+
+## 12.0.0-next.6
+
+## 12.0.0-next.5
+
+## 12.0.0-next.4
+
+## 12.0.0-next.3
+
+### Minor Changes
+
+- 39daa92: Add new grain and organic icons
+
+## 12.0.0-next.2
+
+## 12.0.0-next.1
+
+### Patch Changes
+
+- 7f0d476: Rename disableGlow into showGlow
+
+## 12.0.0-next.0
+
+### Major Changes
+
+- 60e73a2: disable panel glow by default. showGlow is true by default.
+
 ## 11.19.0
 
 ## 11.18.1